requirements.txt
torch>=2.0
torchvision
numpy
opencv-python
matplotlib
pandas
scikit-learn
seaborn
pyyaml
tqdm
Pillow
tk
torchvision
ultralytics

